Fix to uint256

+++ contracts/libraries/VaultLifecycle.sol
@@ -27,7 +27,7 @@ library VaultLifecycle {
         address USDC;
         address currentOption;
         uint256 delay;
-        uint16 lastStrikeOverride;
+        uint256 lastStrikeOverride;
         uint256 overriddenStrikePrice;
     }
