+++ contracts/vaults/RibbonThetaVault.sol
@@ -168,7 +168,7 @@ contract RibbonThetaVault is RibbonVault, RibbonThetaVaultStorage {
      * @notice Sets the new discount on premiums for options we are selling
      * @param newPremiumDiscount is the premium discount
      */
-    function setPremiumDiscount(uint16 newPremiumDiscount) external onlyOwner {
+    function setPremiumDiscount(uint256 newPremiumDiscount) external onlyOwner {
         require(
             newPremiumDiscount > 0 &&
                 newPremiumDiscount < 100 * Vault.PREMIUM_DISCOUNT_DECIMALS,
